refactor fs master test
